pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"


[project.scripts]
open-api-maker = "open_api_maker_kamaalio:cli.cli"


[project]
name = "open_api_maker_kamaalio"
version = "0.0.2"
authors = [{ name = "Kamaal Farah", email = "kamaal.f1@gmail.com" }]
description = "A package that transforms Swagger to OpenAPI."
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = ["PyYAML >= 6.0.1"]


[project.urls]
"Homepage" = "https://github.com/kamaal111/open-api-maker"
"Bug Tracker" = "https://github.com/kamaal111/open-api-maker/issues"
